try to fix unsigned range issue

+++ pulse2percept/models/_nanduri2012.pyx
@@ -55,7 +55,7 @@ cpdef spatial_fast(const float32[:, ::1] stim,
 
     """
     cdef:
-        size_t idx_el, idx_time, idx_space, idx_bright
+        int32 idx_el, idx_time, idx_space, idx_bright
         uint32 n_el, n_time, n_space, n_bright
         float32[:, ::1] bright
         float32 px_bright, denom, d2c, d2e, amp
@@ -158,7 +158,7 @@ cpdef temporal_fast(const float32[:, ::1] stim,
         float32 t_sim, amp, scale
         float32[:, ::1] all_r3
         float32[:, ::1] percept
-        size_t idx_space, idx_sim, idx_stim, idx_frame
+        int32 idx_space, idx_sim, idx_stim, idx_frame
         uint32 n_space, n_stim, n_percept, n_sim
 
     # Note that eps must be divided by 1000, because the original model was fit
